Fix 56 non-matching functions via tooling fixes + codegen corrections
Verification improvements (tools/):
- verify_all.py: Fix relocation type 312 (R_AARCH64_LD64_GOT_LO12_NC)
  misclassified as ADD_LO12 — should be LDST64 (needs >>3 shift). +17 verified.
- fix_movz_to_orr.py: Extend MOVK lookahead from 1 to 4 instructions to
  handle interleaved register computations. +1 verified.

Source fixes (src/app/):
- Fix 11 PTR_DAT addresses off by +0x1000 (systematic Ghidra extraction error)
  across fun_medium_{014,017,018,021,022}.cpp. +11 verified.
- Add __attribute__((noinline)) to prevent unwanted inlining of
  FUN_710014c900 and FUN_71001793a0 in same-TU callers. +3 verified.
- Change return type u64→u32 for 18 functions returning 32-bit values
  (mov w0 vs mov x0 codegen). +18 verified.
- Add __attribute__((visibility("hidden"))) to 30 extern symbols that use
  direct adrp+add in the original (not GOT adrp+ldr). +33 verified.

Net result: 5877 → 5933 verified (+56 functions, +0.7%)

+++ src/app/fun_batch_c_010.cpp
@@ -282,7 +282,7 @@ u64 FUN_71021f6514(s64 *param_1)
 }
 
 // 0x71021f72f0
-u64 FUN_71021f72f0(void)
+u32 FUN_71021f72f0(void)
 {
     FUN_71015c0410();
     return 0;
@@ -313,21 +313,21 @@ u64 FUN_71021fc278(s64 param_1, s64 *param_2)
 }
 
 // 0x71021ff030
-u64 FUN_71021ff030(lua_State *param_1)
+u32 FUN_71021ff030(lua_State *param_1)
 {
     app::item::apply_powerup_throw(param_1);
     return 0;
 }
 
 // 0x71021ff1c0
-u64 FUN_71021ff1c0(lua_State *param_1)
+u32 FUN_71021ff1c0(lua_State *param_1)
 {
     app::item::apply_powerup_etc(param_1);
     return 0;
 }
 
 // 0x71021ff230
-u64 FUN_71021ff230(lua_State *param_1)
+u32 FUN_71021ff230(lua_State *param_1)
 {
     app::item::apply_powerup_attr(param_1);
     return 0;
@@ -358,7 +358,7 @@ u64 FUN_7102208ccc(s64 param_1, s64 param_2)
 }
 
 // 0x7102208d00
-u64 FUN_7102208d00(lua_State *param_1)
+u32 FUN_7102208d00(lua_State *param_1)
 {
     app::boss_private::send_event_on_boss_dead(param_1);
     return 0;
@@ -374,7 +374,7 @@ u64 FUN_7102208df0(s64 param_1)
 }
 
 // 0x710220bf50
-u64 FUN_710220bf50(lua_State *param_1)
+u32 FUN_710220bf50(lua_State *param_1)
 {
     app::item_collision_manager::disable_ground_collision(param_1);
     return 0;
